Split wrapper argument tests
Move the argument helper tests into a dedicated submodule so
arguments.rs stays under the 400 line limit while keeping the
test coverage intact.
